Don't let the IP_address class export ip as a feature.
Exporting the IP as a feature introduces the need of refreshing the
IP of all the Hosts using that instance of IP_address. This commit
also adds a check on the base Node.py functions so that if class has
no features it is not possible to open the "Edit" dialog.

WARN: This commit changes the semantic of the IP_address class.
      Only Hosts will be able to change the IP of the IP_address
      instance. What happens if more than one Host use the same
      IP_address instance? Is this possible?