Host.py: fix mistake in refreshing the interface data
If the set interface is non-existent (e.g. the given interface name is
not valid on the machine where Locator is running), the refresh_data.sh
script will return an empty string, which was not correctly handled
by the method of the Host class. This patch lets the handler return
immediately if an empty string is returned.

Reported-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>