Introduce dependencies in node connection mechanism.
This patch lets connections always be created using dependencies,
that is, if a subclass of Node needs an object of another subclass
of Node to be correctly instantiated, that other object comes first
in the connections list. E.g. if a WebServer needs a Host, which
in turn needs an IP_address, the correct sequence of connections
will be: IP_address - Host, Host - WebServer.
This mechanism prepares the code to introduce an object relational
mapping layer. The ORM will need to get an object when all its
dependencies are satisfied.
NOTE: this patch leaves a debug print in locator.py; this also fixes
breakage of the Host creation button due to the new functionalities
introduced in previous patches.