Add demonstrative code for the usage of sqlalchemy.
This patch expands the save() and load() methods of the Locator class,
letting the user save a connection list in a sqlite database with a
simple click. This won't change anything in the internal data structures
of the current Locator class, but will just save the state in a SQL
database and print on the terminal all the operations. This prepares the
code to be extended in a more intelligent and constructive way.