locator.py: let connections disappear along with nodes.
Unfortunately, the queue_draw() method periodically invoked by the
timer does not clear connections, as that area of the window does
not show as "modified" in the gtk internal data structures. After
destroying a node and erasing its connections from the Locator data
structures, the main window must be cleared and everything else
left must be redrawn.

Reported-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Serena Ziviani <senseriumi@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>